feat(construction): Add components for applying Constructions
This commit adds components that allow users to apply Constructions and ConstructionSets directly to honeybee objects. These are meant to parallel the legacy components for applying constructions but, given that ConstructionSets effectively allow users to specify constructions by face and object type, these components are free from having to organize the input constructions this way.  For example, the "Apply Opaque Construction" component can be used to apply a construction to a Face, Door, or Room and we aren't limited to only using it for Zones as we were in legacy. For the case of a Room, I have also included the ability to assign different constructions based on cardinal orientation, which is a workflow that I think many will appreciate.  Because of this, I feel more comfortable removing the ep_constr_ input from the "Glazing by Ratio" and "Louver Shades" components in honeybee-grasshopper-core (something I know Mostapha wanted to do). So I think I will remove those inputs once we merge this in.

There are also a few other tweaks I made to the other construction components, which are all about the docstrings and organization in tabs.