feat(construction): Add WindowConstructionShade Component
This PR is coordianted with [this one over on honeybee-energy](https://github.com/ladybug-tools/honeybee-energy/pull/211).

It fixes the way that blinds and shades were represented in the schema previously, which could never be simulated by EnergyPlus since E+ always needs a ShadingControl object in order to be able to simulate them. So I added a new WindowConstructionShaded class to handle the creation of this ShadingControl object and allow users to change some of its properties.