fix(hvac): Remove VAV templates with no Reheat
Not only are these systems not working now but we figured that they're just not worth supporting since the case that they're relevant is so rare.  And, for the handful of cases where they are relevant, their performance is virtually equivalent to the systems that have the reheat.  So or response to these handful of cases is "just use the template with the reheat and you won't see the reheat turn on in the simulation if it is truly a relevant case."