fix(docs): Remove non-ASCII characters from docstrings
These non-ASCII characters mess with the JSON serialization and they really are not necessary.

This commit also adds JSON versions of all components. These won't be included in the Python package but they are meant to be updated as the user objects and source are updated. This way, other developers who want to reimplement the Ladybug Tools plugin in other visual scripting interfaces have a data for a good starting point.