feat(simulate): Add component for computing annual loads
For now, I am replacing the "HB Model to IDF" component with this "HB Annual Loads" component since it was kinda redundant with the OpenStudio component.  Now, there's a good use case for the speed of that component's fast serialization to IDF since it means we can get the annual loads of a shoe box in under 2s.