feat(result): Add component to separate data by object type
The new component can also normalize the data by the area of faces/apertures.

Now that this new component exists, I am also taking the opportunity to decrease the size of the "HB Read Face Result" since the new component can be used to separate the data and the other inputs are redundant.

I'm also improving the docstrings on a number of components and simplifying the code that does floor area normalization.