fix(simulate): Add a check for the most intense design day in Peak Loads
I noticed that some of the recent OneBuilding .ddy files are using the recommended practice of including a cooling design calculation in each month instead of only providing a single Summer design day. This is great but the peak loads component was not set up to take advantage of this. Now the component will run an initial calculation to find the most intense cooling design day before producing final results.