fix(simulate): Improve shade benefit calculation so the result is valid
It seems there was no good reason to multiply by the area of the windows. Just dividing by the area of the shade cells was sufficient. Also, I realized that it's better practice to simulate all of the apertures of a face together since shading of separate apertures on the same face are likely to interact with one another.

Lastly, I added a lag_time input to offset the time of cooling load from the time of solar load. The result that it produces seems to make more sense and I intuitively know that a lag time of 1 hour is a reasonable default from the simulations that I have run.